+++ src/stores/preferences.ts
@@ -27,7 +27,6 @@ export type UserPreferences = {
 	useNewPollInPollist: boolean
 	useCommentsAlternativeStyling: boolean
 	useAlternativeStyling: boolean
-	useCardsArrangement: boolean
 }
 
 export type SessionSettings = {
@@ -65,7 +64,6 @@ export const usePreferencesStore = defineStore('preferences', {
 			useNewPollInPollist: false,
 			useCommentsAlternativeStyling: false,
 			useAlternativeStyling: false,
-			useCardsArrangement: false,
 		},
 		session: {
 			manualViewDatePoll: '',